Added test for grade and rate.

+++ src/main/java/gradebook/Rate.java
@@ -1,5 +1,5 @@
 package gradebook;
 
-public record Rate(int value) {
+public record Rate(double value) {
 
 }
